merge adjacent free blocks
TODO:

When allocating memory from the free list, the remainder of the block
must be calculated.

If the remainder is larger than the minimum required size, then the
remainder must remain in the list.

This amounts to rewriting the FreeBlock element at its new position,
and setting the size to the remainder.